Added Log Nutrition
Basic Add nutrition, I fixed registration and it brings back to the registration page. 
After registration it asks the user to login again to make sure user is authenticated. Didnt push to master since I wanted to discuss. 
Looking to initiate some views, not sure where the database is located.