Save/Get is simpler, Schedule/Routine are almost done
Class based is easy now we should be able to edit the forms/models appropriately and create a meaningful chart in profile. 
Disabled some CSS for radio buttons since they deemed my selectors invisible. 
Added button to view what is in the database which does it through a simple get request.
Did Schedule and routine skeleton though values are not that meaningful. We should be able to make them so.
Will work on this Thursday after work and possibly finalize it. Let me know if you can come up with a meaninful schema. 
I possibly can spend Friday night to Saturday morning to see if Django can magically make recorded data into charts and such.